Lecture-Management-Public/requirements.txt
flask_login==0.6.2
flask_wtf==1.1.1
mysql-connector-python
openpyxl==3.0.10
python-dotenv
WTForms==3.0.1
gunicorn==20.1.0
flask==2.2.2
flask-sqlalchemy==3.0.2